Trunk: Fixed a gcc compilation error related to rev 1000
git-svn-id: svn://localhost/public/MoSync/trunk@1004 6ad7a17e-d552-4a1e-ace4-f958fe14b117

+++ tools/pipe-tool/Symbols.c
@@ -1611,7 +1611,7 @@ void DumpMetaData ( FILE *out )
 			fprintf(out, "<%s,",Sym->Name);
 			fprintf(out, "%s,",Hex32(Sym->Value));
 			fprintf(out, "%s,",Hex32(Sym->EndIP));
-			fprintf(out, "%s",returnType[Sym->RetType]);
+			fprintf(out, "%s",returnType[(int)(Sym->RetType)]);
 
 			fprintf(out, ">\n");
